Use correct URI scheme when opening workspaces (#10092)
Addresses https://github.com/posit-dev/positron/issues/9488

This PR updates what we do here on the Positron side when we are opening
a workspace so it works both on Positron web and desktop. We need to get
the right schema from the path service to do so.

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- Workbench: fixed a bug when opening workspaces such as with
`usethis::create_from_github()`


### QA Notes

If you have a real folder that really exists at a
`/real/path/that/does/exist`, then with this PR you should be able to
successfully run `.ps.ui.openWorkspace("/real/path/that/does/exist",
newSession = FALSE)` from the R console both:

- in a regular desktop build (this always worked)
- in a web build (this is now fixed with this PR)