remote ssh: add custom serverInstallPath setting (#10017)
Addresses #9013. Adds a `remoteSSH.serverInstallPath` setting that lets
you customize the Positron server data directory, instead of the default
`~/.positron-server`.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- The new `remoteSSH.serverInstallPath` setting lets you customize the
directory for the Positron server #9013

#### Bug Fixes

- N/A


### QA Notes

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->
The new setting should be respected. To test on this branch, you'll
unfortunately need to build REH on the remote host and put it where it
needs to be. It'll be easier to test on a daily build.