E2E test: handling for soft fail tests (#10089)
With this PR, you can tag tests that you do not want to see a failure
for in Github with SOFT-FAIL.

For ubuntu, windows, and RHEL the playwright exit code is now ignored
and the JSON results are parsed instead. Failed tests tagged with
SOFT-FAIL do not fail the run (for merge to main).

### QA Notes

@:web