Improve view mode keybinding for markdown cells (#10069)
Related to #9713 (the ask to have a mechanism to switch between modes
already exists)

This PR is a suggestion for improving the `edit` to `view` mode
experience for markdown cells.

While playing around with notebooks, I noticed that there was some
inconsistency with how markdown cells behaved when using the keyboard vs
the cell action bar that caused me a bit of confusion.

When you are in edit mode for a markdown cell, the cell action bar has a
chevron icon button that hides the editor for markdown cells. I assumed
this action was changing the cell from `edit` to `view` mode. The
notebook also has a keybinding to exit edit mode that maps to `Escape`.
This command does not do the same thing as the button in the cell action
bar. When you press `Escape` in a markdown cell, focus moves to the
markdown cell container and leaves the editor. The markdown editor view
stays around with the preview underneath it.

It was unclear that this mismatch in behavior was intentional! I think
another reason that the `edit` to `view` mode transition in markdown
cells is confusing has to do with not having a great visual indicator
for being in `view` mode. We show a focus indicator around the cell when
a user is in `edit` mode, but there isn't anything like that for `view`
mode.

I can close out this PR If we don't want to consolidate the behavior and
need it to be different, but I figured I'd see what folks thought. I
find the consolidation makes things less confusing.

It turns out we have 2 different commands that address exiting `edit`
mode:
- `positronNotebook.cell.collapseMarkdownEditor` -> command for action
that shows up in markdown cell action bars
- `positronNotebook.cell.quitEdit` -> command to exit `edit` mode for
all cell types that is not contributed to any action bars

I think we still need two separate actions, one that allows us to
contribute a button to a cell action bar, and a generic one that we can
attach the keybinding to. I just think the behavior for both should be
the same!

**BEFORE**


https://github.com/user-attachments/assets/9ac239f8-85ba-4941-9c4f-e3e6d4e89a99


**AFTER**


https://github.com/user-attachments/assets/ab87a398-1717-44de-b8f3-a927d4f18645


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->
@:positron-notebooks

---------

Signed-off-by: Dhruvi Sompura <dhruvi.sompura@posit.co>
Co-authored-by: Wasim Lorgat <mwlorgat@gmail.com>