python: make pet default locator (#10143)
I would like to wait to merge this until after #9186 has been verified
out of an abundance of caution! But queueing it up for now.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- #6582 

#### Bug Fixes

- N/A


### QA Notes

This should be okay with green CI after #9186 has been verified, since
that was the issue that caused us to revert this originally.