Fix how we track foreground session (#10159)
Addresses #7861

This PR makes two changes:

- Updates the `foregroundSession` setter to not reset the foreground
session if we have the same session. This is to address how @wch points
out we have been firing the `onDidChangeForegroundSession` callback
every time you click on the tab list title.
- Switches `doStartRuntimeSession()` to use the `foregroundSession`
setter and not to update the private `_foregroundSession`. I believe
this is the main source of the bad behavior that @wch observed.

This PR DOES NOT address
https://github.com/posit-dev/positron/issues/7052; if you reload the
window with an R session as the foreground session and another session
(like a Python session), you will not necessarily get that foreground
session back on top after the reload. What this PR fixes is what you get
when an extension calls `positron.runtime.getForegroundSession()`; it
will now report the right thing.

@:sessions

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Fixed what Positron reports to extensions in
`positron.runtime.getForegroundSession()`


### QA Notes

I put some debug logging in the zed extension in
0a653de8097cf2da82497dc1f425309ef2d37a1f and you can use that in a dev
build to see what we're reporting, but I imagine I'll revert that commit
after we confirm this is doing the right thing. Probably the easiest way
to confirm what's going on in a release build is to look at what
Positron Assistant reports as the current session as @wch points out in
https://github.com/posit-dev/positron/issues/7052#issuecomment-3099599576:

- Start an R session and Python session
- Select the R session
- Reload Positron
- Look at the Positron Assistant chat; it should always report the
correct foreground session (maybe you'll see R as foreground briefly
before Python takes over)
- Select the Python session
- Reload Positron
- Look at the Assistant chat and confirm it reports the correct
foreground session as sessions start up

Do be aware of #7052; we do not persist the R foreground session
correctly across a reload.