Dynamically fetch Bedrock models (#10131)
Address #9738 #9643

Use the Bedrock SDK to query for models. It filters for Anthropic models
that are active and use an inference profile. Then it fetches the
inference profiles to match up the inference ARN for each model.

When making requests to Bedrock, the inference ARN is used to ensure
that the model can be used. This allows Claude Sonnet 4.5 to be used
without inference profile errors.

The listing also includes Sonnet 4.5 and Opus 4.1. Sonnet 3.5 is also
there but because the model status hasn't been marked legacy yet and
Posit still has inference profiles that include it.

The connection test now resolves models to validate the credentials. It
uses the node provider chain to retrieve credentials just like the
Vercel SDK so it's a suitable test.

There is support to add in other providers but this queries for
Anthropic models only.

<img width="239" height="328" alt="image"
src="https://github.com/user-attachments/assets/d046cc57-aef6-4286-94d1-bda938d4065a"
/>


### Release Notes

#### New Features

- Dynamically fetch Bedrock models for Assistant

#### Bug Fixes

- N/A


### QA Notes

---------

Signed-off-by: Tim Mok <timtmok@users.noreply.github.com>
Co-authored-by: sharon <sharon-wang@users.noreply.github.com>