Cleanup notebook errors in dev tools when opening positron notebook (#10197)
Addresses #10022

This PR addresses the sad state of the dev console when opening up a
positron notebook.

Before this PR opening a notebook would cause a cascade of red in the
devtools console...

Before changes:
<img width="814" height="1282" alt="image"
src="https://github.com/user-attachments/assets/fa33db3f-7f9b-4603-a8d4-f92cb6a7ab83"
/>

After this PR it's a much less chaotic scene:

<img width="816" height="276" alt="image"
src="https://github.com/user-attachments/assets/cd0325f5-c569-4e2d-bcee-fcf230bb24f4"
/>

## Fixes

There were three main error fix areas and one general log hygiene fix

1. We were adding a scoped context key service to each cell's monaco
editor widget. The problem was the monaco editor widget itself was
adding a context key service and when it tried to and found there was an
existing one, it errored. No obvious functionality problems, but still
not ideal. Fixed this by just letting it use the one it creates later.

2. The Monaco editor was trying to get the editor progress service from
DI. The problem is this service is not available outside of standalone
full-page monaco editors. We copied the VSCode notebook approach to
fixing this which is to just give it a dummy service when it tries to
import.

3. The editor manager was attempting to access the onDidFocusWidget
property on the notebook instance. This is one of the properties we
hadnt yet implemented from INotebookEditor. Fix was just adding it and
hooking it up properly.

4. We were really overzealous with our logging in the Info level. I
switched most of the logs to the debug type to avoid polluting the
console.


### QA Notes:

There should be no behavior changes from these updates. (Maybe a bit
more stability?)

The logs should be a lot cleaner.