catalog explorer: more informative error handling (#10127)
Adding in a small logger + more informative toast messages. Addresses
#10108 since the root of the errors was the fact I had a handful of
dummy Databricks accounts (eg, `https://abc`). We should only allow
users to create Databricks accounts that exist and can be connected.

Additionally, I'm sneaking in a logger to help out with user error
handling/understanding of processes.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- #10108 Validate Databricks connections before creating them.


### QA Notes

- enable catalog explorer
- configure a databricks catalog provider with URL `https://abcde` and
PAT `dapi123`
- see `fetch error` on creation
- also make sure we can still connect to real accounts
- look at `Output` > `Catalog Explorer` for logs