python: skip failing venv tests (#10187)
The failing `Test TypeScript (macos-latest, 3.x, venv)` runs are failing
after #10143 merge. These tests are integration tests that listen for
events from the `js` locator; specifically, it is failing on tests that
wait for events emitted by the `FSWatchingLocator`. They started failing
since that is no longer the default.

Some options for later
- updating tests to be locator agnostic events
- trying to force a different setting in the integration tests, at some
point before the test runner starts up. I thought we'd be able to switch
the locator just for the test, but we'll need to restart Positron, which
is a bit messy in this integration test. You can see this attempt in
[bd3dae8](https://github.com/posit-dev/positron/pull/10187/commits/bd3dae8fdcc401c8da2870a667445efde654d412).

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- 


### QA Notes

green ci