Add support for WSL remotes (#10091)
This change adds a built-in WSL remote extension to Positron for
Windows. WSL is now a second remote destination alongside SSH on Windows
machines:

<img width="1024" height="814" alt="image"
src="https://github.com/user-attachments/assets/ac9716f3-0e0b-42ab-a631-5f3c5abc128d"
/>

It is a hard fork of
[open-remote-wsl](https://github.com/jeanp413/open-remote-wsl), and
follows exactly the same playbook (copied with MIT notifications and
attribution; reformatted; retrofitted to build as a built-in extension).

This change also addresses what remains of
https://github.com/posit-dev/positron/issues/9047. After some
investigation, it appears that the problem with domain sockets in WSL
didn't have anything to do with WSL but with a bug in the `request`
package that caused domain socket requests to try to use a proxy. That
problem was fixed by removing `request` and replacing it with Axios in
#9421.

There _was_ one issue in the logs for #9407 that could have caused
trouble; it appears that in some circumstances the supervisor could try
to create its domain socket in a folder that didn't exist. I fixed that
in https://github.com/posit-dev/kallichore/pull/52, and the new
supervisor version with the fix is included in this PR.

Addresses https://github.com/posit-dev/positron/issues/9047
Addresses https://github.com/posit-dev/positron/issues/4366


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Add native support for remote sessions on WSL (#4366)

#### Bug Fixes

- Fix errors connecting to WSL when using IPC transport (#9047)


### QA Notes

- This feature will be difficult to test until it is in a daily build
since it relies on downloading a matching Positron binary.
- WSL works almost exactly like SSH and you should expect the same level
of functionality.
- The WSL feature should not show up on non-Windows environments.
- The support matrix for Linux WSL hosts is the same as the support
matrix for Linux hosts in general.