catalog explorer: rename commands (#10203)
Updates 

- `Copy {R/Python} Code` -> `Preview Table via {R/Python}` and `Download
File via {R/Python}`
- `Open in Session` -> `Preview Table in Console` and `Download File in
Console`

Also added a notification for user feedback that code has been copied to
clipboard.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- #10107 Updates command names to copy code from the catalog explorer.

#### Bug Fixes

- N/A


### QA Notes

- enable `catalogExplorer.viewTestCatalog` and `catalogExplorer.enable`
- add demo catalog
- right click on table, see the updated command names
- right click on csv, see the updated command names