Add R workspace save/restore setting (#10213)
- Addresses https://github.com/posit-dev/positron/issues/964
- Closes https://github.com/posit-dev/positron/pull/9936

This is a non-fork PR because of the problems with our CI outlined in
https://github.com/posit-dev/positron/issues/6628, although I made some
changes to make _one_ setting instead of having two separate settings
for saving and restoring. I wholeheartedly agree with @jennybc in
https://github.com/posit-dev/positron/issues/964 that it doesn't make
sense to have separate settings. We are not going to address the "ask"
option currently.

We have the needed changes to ark in
https://github.com/posit-dev/positron/pull/10199.

### Release Notes

#### New Features

- R: added new setting `positron.r.saveAndRestoreWorkspace` to opt in to
saving a workspace to `.RData` on exit and restoring on startup.

#### Bug Fixes

- N/A


### QA Notes

It would be great to verify this on a Windows machine because we had a
bug in this behavior on Windows, and verification likely should happen
together with https://github.com/posit-dev/positron/issues/9927.

- Start an R session, and define some variables.

```r
foo <- 1
bar <- sample(letters, 3)
```

- Restart R, and the variables are gone
- Change `positron.r.saveAndRestoreWorkspace` to true and restart
Positron
- Define the variables again, then restart R, and notice that the
variables are BACK, BABY
- Notice that you have a new file in your workspace called `.RData`
(this is the file that can be restored from)
- Change `positron.r.saveAndRestoreWorkspace` back to the default and
restart Positron
- When R comes back up, confirm there are no variables (this validates
https://github.com/posit-dev/positron/issues/9927 is fixed) even though
you have an `.RData` there

---------

Co-authored-by: Dianyi Yang <dianyi.yang@politics.ox.ac.uk>
Co-authored-by: Jenny Bryan <jenny.f.bryan@gmail.com>