Fix leaking supervisors on Remote SSH (#10147)
This change addresses an issue in which the feature that is intended to
reattach to long-running sessions does not work on Remote SSH, with the
result that a remote host can accumulate a large number of orphaned
sessions.

The primary issue is a regression from the fix for
https://github.com/posit-dev/positron/issues/6174. In that change, we
started the supervisor early (before Positron itself started). This
change was only _intended_ for Positron Workbench/Web, but it also
affected Remote SSH since the these configurations share a lot of code,
with the result that we were starting a supervisor before we got a
chance to reattach to one.

There was a secondary issue, also addressed here, in which we would
forget your sessions and/or their contents if you didn't have a
workspace open, because all of the storage systems use workspace storage
scope. These have been addressed by using profile-scoped storage for
these situations.

Finally, one thing that makes this situation challenging to debug is
that just looking at the process list doesn't tell you anything useful
about what all those `kcserver` processes are -- they could be leaked
from days or weeks ago (hopefully not!) or they could be from workspaces
you legitimately intend to return to. So I've added a new _Kernel
Supervisor: Show Running Kernel Supervisors_ command.

<img width="623" height="193" alt="image"
src="https://github.com/user-attachments/assets/369adb72-4152-4c00-ba54-67ff81baedb2"
/>

This opens a quickpick that shows you all the running supervisors on
your machine. Each one is annotated with the workspace it's associated
with and some other useful info. Selecting a supervisor will give you
the opportunity to see what it's running, open the associated workspace
(thereby reattaching to the supervisor), shut down the supervisor, etc.

<img width="618" height="272" alt="image"
src="https://github.com/user-attachments/assets/06e55560-7f8b-4ce9-984e-57a081d9e488"
/>

It is my hope that most users will never need to know about or use this
tool; it is intended to inspect the state of the system as a dev/QA
troubleshooting aid rather than to be part of end-user workflow.

Addresses https://github.com/posit-dev/positron/issues/6692
Partially addresses https://github.com/posit-dev/positron/issues/5982

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Address issue that could create orphaned R and Pythons sessions in
Remote SSH (#6692)


### QA Notes

You need to use Remote SSH to fully validate this, but note that the
"idle timeout" feature that allows you to close Positron without ending
your sessions also applies to the regular desktop version. So you can
try something like this:

- set _Kernel Supervisor: Shutdown Timeout_ to 4
- close and restart Positron
- open 2-3 windows and start R and Python sessions in each one
- in one of the windows, run _Kernel Supervisor: Show Running Kernel
Supervisors_