Fix Positron Assistant startup on Windows (#10256)
This change fixes an issue that prevents Assistant from starting up on
Windows. The main problem is here:


https://github.com/posit-dev/positron/blob/d3e82f32b5e6a763a226e915ea5732fdc3ee78d5/extensions/positron-assistant/src/promptRender.ts#L287-L292

This regex doesn't work on Windows since Windows uses `\r\n` for line
endings, which means that
- yaml headers are not extracted, so
- yaml headers can't be parsed, so
- we can't identify which commands any of the metadata files are for, so
- no metadata is found for commands, so 
- an exception is thrown during the `activate()` method, so
- the Positron Assistant extension fails to activate, so
- the Copilot Chat extension (which lists Positron Assistant as a
dependency) fails to activate so,
- no chat features work at all.

The only vital part of this change is updating the regex so it works on
Windows. But I've also made a few changes to try to make this slightly
more fault tolerant:
- if we fail to register a command, we log an error but don't fail
activating the whole extension
- at the top level we again wrap in a try/catch; any errors here are
converted to notifications but again do not block activation entirely
- more detailed logging when reading command metadata (where we were
looking, what commands we did find, etc.)

Addresses #10233

---------

Signed-off-by: Jonathan <jonathan@posit.co>
Co-authored-by: sharon <sharon-wang@users.noreply.github.com>