e2e: Skip assistant tests temporarily (#10268)
The echo provider appears broken, which is used by almost all the assistant tests.  Since it's a dummy provider, it's not directly related to an issue with real providers.

To get rid of the noise, skipping them all for now. 

They'll get turned back on once we investigate and determine the cause for echo not working.

### QA Notes

<!--
  Positron team members: please add relevant e2e test tags, so the tests can be
  run when you open this pull request.

  - Instructions: https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
  - Available tags: https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->