Fix regression in `executeCode` actions (#10283)
Related to https://github.com/posit-dev/positron/pull/10191

Well, I learned something today! 🙃 The helper function that I created in
#10191 took the services accessor as an argument, but that is NOT
ALLOWED. You can't get the services correctly after an async operation,
because of how the accessor is designed to manage security and
dependency management. This PR updates how the code is structured to get
the services all at once and pass _them_ instead of the accessor.

The broken tests that caught this are tagged @:editor

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

With this change:

- the new functionality from
https://github.com/posit-dev/positron/pull/10191 works
- the old functionality for
`workbench.action.positronConsole.executeCode` and
`workbench.action.positronConsole.executeCodeWithoutAdvancing` also
still work